Remove migrations
We do not use migrations in deployment tools for now. In order to make
administering development a bit less of a pain, we removed existing
migrations and put migrations folder on ignore list.

When application is deployed, `python manage.py makemigratons
cfy_wrapper` needs to be run before database is set up. This is already
done for you if you use `run.sh` script to reset the database and setup
initial user credentials, so there should be n major functionality loss.

Change-Id: Iadebb4f6bcd85a0ccbad1536244aec3edf3f5473