Refactor blueprint model
Blueprint is now always represented as a folder on deployment service.
We do our best to validate it, but there are still scenarios where user
can upload invalid blueprint. Those failures will be caught later on
when we publish archive to Cloudify manager.

Folder management of the blueprint has been delegated to signals that
can reliably detect changes in blueprint model.

Change-Id: Ic8b7f9b2f13ee358d19752053e5a4f55cfee19a0