Synchronize views with model and task changes
Since most of the internal workings of the deployment service went
through some rather profound changes, views needed some care to get them
back in sync.

Note that some of the existing views have been removed. All of the
functionality of the old views can be simulated using new set of views
that are cleaner and easier to use. If the need arises, they can be
added back.

These changes probably lest CLI tool unusable. This should be remedied
as soon as possible.

Change-Id: I2cda590c58117ef1b0a3bbb138877ea2ac49bc27