Refactor container model
Container serves as a central synchronization object in deployment
service. There can be at most one task being executed on container at
any time. This can be achieved by atomically flipping the busy field on
container.

If busy field is set to true, no modifications should be done to
container. If busy field is set to false, it must first be set to true
and saved. If this operation fails, again no modifications should be
done to container, since someone already gained access to container
before we did. After we successfully set the busy flag, we are free to
modify container in any way we like.  After we are done, busy flag
should be set to false.

Change-Id: I05c448ef6a69ca82842bc7ddc47a4b729fd8751d