Add OpenAPI documentation
In this commit, we replaced inline API documentation with separate
schema file.

In some regards this is a step back, since documentation is now in a
separate place from implementation that realizes promised schema.

On the other hand, by having separate schema, we gain a few benefits
too:

 1. Documentation can be edited in YAML aware editor. Editing inline doc
    strings formatted as YAML can be tedious. Swagger even offers free
    online editor for people who like that kind of stuff.
 2. Documentation can be DRY-er, because we can factor out common pieces
    of schema and reuse them.
 3. Having a complete OpenAPI schema available opens new interesting
    ways of interacting with our api. Example can be seen at [1].

[1] http://www.coreapi.org//tools-and-resources/example-services/

Hopefully this justifies the move of documentation to external file.

Change-Id: Id0d6fbb30c863bcfaebd8867bfd3befa2909949c