Add busy field to container model
Determining when the operation finished from blueprint's status is
feasible, but really annoying task that is quite error prone. In order
to reduce clients' burden a bit, container now directly signals when it
is in idle state.

Note that this change imposed no changes of server's inner workings,
since we already use this piece of information. We only exposed it to
the client in convenient form.

Change-Id: I0a74e0be4c8b3a31cf78996b135ccd89838bdfab