Restore access to container nodes
During the transition to new tasks, access to all compute nodes within a
selected container has been lost. This commit restores functionality and
adds tests that were non-existent before.

This commit should make it possible to have fully functional console
client with all of the functionality of the pre-refactor service.

Change-Id: I74421cdb9eec004d97b56a18444631f159260d42