Add new console client for deployment service
We added new console client for DICE deployment service. Main reason for
addition is move from bash to python, since there are some things that
just cannot be done with reasonable amount of effort from bash.

Concrete example of such a thing is getting sensible behavior from curl
when request fails. We can either get complete JSON message that server
returned and exit status of 0 or we can get sensible exit status from
curl, but we loose the server message.

We might be able to hack together some ugly grep or sed script to
extract information from JSON, but debugging such a script would be
nightmare.

Doing exact same things in python simply bypasses all of the troubles
listed above, since we can easily parse JSON and then use language
features to deconstruct the message.

New script should be completely backwards compatible. We kept all of the
operations that still make sense named exactly the same as in the old
`dice-deploy-cli` script. Outputs should also be compatible. There might
be a few operations missing, but they are not present manly because
back-end does not support them yet in the same form as before.

One set of missing commands is the one related to creation and deletion
of backend inputs. This set of commands has been replaced by single
`set-inputs` command that should make life a bit easier for the admins.

Change-Id: I12e6e5f85ce54dabe2d6c80ad17d729fcdb28840