Update blueprint input handling
We finally managed to get deployment service mature enough to support
deploying blueprints without internally mangling them. This means that
we forward blueprints with no modifications to Cloudify manager.

In order to be more user friendly, we added extra step in deployment
creation that checks for presence of required inputs and prepares
sensible error message for end user.

Change-Id: I959ad76baf7d53c3f6cf45763039432093ddc230