Make validation errors more informative
New validation logic should make it clear what exactly went wrong when
blueprint has been uploaded. Note that the check is not exhaustive,
because we like to return result to user as soon as possible.

More strict checking is done asynchronously in celery worker. Results
of those operations can be obtained by querying api.

Doing this simple cleanup uncovered a bug in test suite that was also
fixed by this commit.

Change-Id: Ib2f04d5fcfb034305458b790d4f06a3418430bf7