Raise upload limit (again)
We underestimated the sizes of jar files that real world applications
come in. Since most of them seem to package quite a lot of data, we
raised the limit to roughly 2 GB. We are now pretty sure jvm will fail
to run the application well before we hit our size limit.

Change-Id: I2ff14599f50c680365a4568142da92b25afa7786