Allow storing empty string as an input value
There are valid scenarios where user would like to have an input whose
value is an empty string. By default, this is impossible to enforce in
django models, because any missing text values are stored as empty
strings in database.

To get around this problem, we modified the string storage for input
values. It is now possible to differentiate empty string from None.
Hopefully, this rather hackish solution will not come back to haunt us
in the future.

Change-Id: I0dad73b02e1ebb6e276257ce89f16ff301fb3b82