Enable RabbitMQ web UI on development machine
Debugging queue functionality of the DICE Deployment Service is quite
painful without an UI to look at. We can use CLI interface to manage
RabbitMQ, but web UI is far nicer and faster.

This change enables management plugin in RabbitMQ and adds a port
forward rule to Vagrant file so we do not need to jump through hoops
to get to the UI.

Change-Id: Id82d809841dbfe79339be1f433eff76f56a24f47