Split deployment creation into two tasks
This split is needed in order to be able to retry certain operations
when their failure is caused by connectivity problems.

For example, if the connectivity error occurs while the task is
waiting for deployment to initialize, we would like to restart status
polling without recreating the deployment. Now that those two
operations live in separate tasks, this is trivially achievable.

Change-Id: Ib89eb51f99e4727a0b7f81d2f52c3edad796a2cc