Refactor installation task
Blueprint installation task is now only responsible for scheduling
installation. Waiting for completion is done by separate tasks whose
only job is to monitor execution.

Task that waits for execution expects execution id to be returned by
the task preceding it. This should make it easier to reuse in the
future, since waiting for certain operation to finish can be
implemented by simply returning proper id and chaining waiting task.

Change-Id: I4d8b1a5e43d9aa5b8af57a725f871485f06b24df