Handle secure Cloudify manager
When connection to Cloudify manager over https, some additional
information needs to be passed to client creation function. This
commit adds two new settings that control the protocol that is used to
connect to manager and server certificate that should be used to
validate connection.

Default settings were selected in such a was that preserve current
behavior of using unsecure connection to manager.

Change-Id: I6258e73725f648d4f8ce47c65abbd8065dc07186