Add instructions for bootstrapping Cloudify on AWS
Instructions for bootstrapping Cloudify Manager on AWS differ from
official documentation in few important aspects.

First major change is that we install manager into dedicated VPC that
is created during bootstrap. This solves majority of the issues with
instances not being able to boot because they can only be used from
VPC. This change does not raise the costs of running a manager, since
creating and running a VPC is free.

Secondly, we only support accessing manager over https. This is
enforced because blueprints that we are sending to the manager can
contain sensitive data and sending it in plain text over the wasteland
that is internet is irresponsible at best.

Last major change is use of customized AWS plugin. We were forced to
introduce this change in order to properly handle preexisting elastic
IP addresses in VPC. This change was needed in order to support using
self-signed certificates on manager.

Change-Id: I97fe0032848174e8c66b7e14033f8eb1fafc284b