Add separate subsection in tool admin doc
The `doc/AdminGuide.md` already had a description of how to
configure the `dice-deployment-cli`, but it was included in a
larger Configuring the service section. Since it is useful to
link to the setting up of the client's CLI from other documents,
a separate heading will help anchor the link while not confusing
end users with the setting up of the service.

Change-Id: Ib6602d897ea07de5495f6dfe287f9723e38e091d