Add metadata upload capability to deploy endpoint
When uploading blueprint, custom metadata can now be attached. This
metadata is passed to the DMon when/if application gets registered.
Metadata is completely free form and we pose no restrictions upon it.

It is end-users responsibility to only supply metadata that DMon
understands. This might change in the future if the metadata format is
defined more strictly.

Change-Id: I6397ead1f2f9cd3a49c9c52e42ab7f38f0074185