Add general instructions for infrastructure
The user experience is much better if the user (the admin)
can prepare in advance what kind of data is needed up front
and what will be generated in the process. We also do not
want the user to revisit the same steps such as clicking through
pages and panels in the FCO web console. Instead, we give
instructions on where the user can get the information on the
target infrastructure that will be useful for configuring
DICE Deployment Service as well as Cloudify Manager.

The `AdminGuide.md` now covers all the relevant information
for Amazon EC2 set-up and for the FCO. The OpenStack information
is only partial currently.

Adding infrastructure specific information to `Cloudify.md`
as well.

Change-Id: I032e9994be543b606a687ee9b1ba5aa75c17e2ff