Update documentation to unify instructions
Now that `Cloudify.md` requires steps done with the DDS source,
and `AdminGuide.md`, conversely, requires some building blocks of
the Cloudify Manager's bootstrapping, we unify the story in
both documents. We now assume that the user will always start
with the `AdminGuide.md`, possibly use `Cloudify.md`, then get
back to `AdminGuide.md`.

For the FCO, this process is now in place.

Change-Id: I39b394dfedafc65e037954194e0d8a7199649f92