Add installation and inputs generation script
The new set of scripts simplify deployment of the DICE Deployment
Service by introducing a single file, `fco-dds-config.inc.sh`
for containing all the configuration values that are needed
in the service deployment and configuration.

The `install-fco.sh` takes care of creating `inputs.yaml` that
is then used in the DDS blueprint deployment process. It also
creates `dds-inputs.json` that can later be used with the
`set-inputs`.

Change-Id: Ifc132ac68278c05b786affeebd6b5639f1a58d6c