Add documentation for new install script
The `AdminGuide.md` now describes the process of deploying the
DDS using the new installation script. The process is now better
streamlined and requires less switching between various configuration
and inputs files.

The generating of the `dds-inputs.json` is now a subject for
developers, because the users will get a pre-generated file.
Therefore the instructions for that are better placed in the
`DevNotes.md`.

Change-Id: I71fd3a438c2e1e5e99fd08e7739a4f816e60a7ea