Add libvirt driver section to Vagrant config
Having Docker and Virtualbox running on the same development machine
can be a bit of a hassle, since we need to manually unload certain
kernel drivers.

In order to bypass this, we can use libvirt Vagrant provider that does
not clash with Docker/Kubernetes. As an additional benefit,
responsiveness seems to improve when Vagrant is using libvirt driver.

Change-Id: I02fcf1311be858126568ebfa090adba8c7c99c56