Toggle default for creating swap on Cfy Mng
Given that our instructions are for VM with 4 GB of RAM, this
falls into the category of "not quite enough of memory". To be
safe, the settings now enable swap to be created on bootstrap
of Cloudify Manager. This should be fine on the AWS as well
given that we create an instance with sufficient hard drive space.

Change-Id: I0b0b0aa9c01654c2e37cbaf271b5fb92b7101467