Add scripts for merging inputs
The new tools help with updating the DDS inputs. Script
`merge-inputs.sh` takes as input two JSON files containing inputs
and produces a merged and updated inputs JSON file. Or, if only
one inputs file is provided, the script fetches the JSON inputs
file to be updated from DICE Deployment Service.

The `merge-inputs.py` is the logic of the merging process, working
just on files.

Change-Id: I2f85eab8a27fbb698827c1e3f69e9879d232a8e6